+++ minesweeper-ui/src/main/java/com/jayasuryat/minesweeperui/grid/MineGrid.kt
@@ -32,6 +32,7 @@ import com.jayasuryat.minesweeperengine.model.block.Position
 import com.jayasuryat.minesweeperengine.model.cell.RawCell
 import com.jayasuryat.minesweeperui.action.MinefieldActionsListener
 import com.jayasuryat.minesweeperui.cell.RawCell
+import com.jayasuryat.minesweeperui.component.InverseClippedBox
 import com.jayasuryat.util.LogCompositions
 import com.jayasuryat.util.dp
 import com.jayasuryat.util.floatValue
